refactor: rename bridge-status utils to fetch-tx-status
